Change the reset vector address to 0x00000000
The specification no longer requires any specific reset vector
address. It is now up to the implementation, so to save space,
the reset vector can safely be moved to 0x00000000.

With this change, the processor now supports the direct exception
vector mode specified in the MTVEC section of the RISC-V
Privileged Architecture Specification.